Separate actions from utils
