+++ ui/src/pages/Withdraw.tsx
@@ -7,7 +7,8 @@ import { useAuthContext } from "../context/AuthContext";
 import { API, getLNURLWithdrawURL } from "../api/api";
 import { Sign } from "../utils/crypto";
 import Input from "../components/Input";
-import { BeautifyNumber, HandleError, NumberRegex } from "../utils/utils";
+import { BeautifyNumber, NumberRegex } from "../utils/utils";
+import { HandleError } from "../utils/actions";
 import Button from "../components/Button";
 import { LNURLEncode, ValidateInvoice } from "../utils/lightning";
 import QRCode from "../components/QRCode";
